+++ App/Containers/RootContainer.tsx
@@ -5,7 +5,8 @@ import {
   StatusBar,
   Platform,
   PermissionsAndroid,
-  Text
+  Text,
+  AsyncStorage
 } from 'react-native'
 import { NavigationContainerComponent } from 'react-navigation'
 import AppNavigation from '../Navigation/AppNavigation'
@@ -62,6 +63,16 @@ class RootContainer extends Component<StateProps & DispatchProps> {
 
   componentDidMount() {
     this.setupLocationMonitoring()
+    AsyncStorage.getAllKeys()
+      .then(keys => {
+        console.log('KEYS:', keys)
+      })
+      .then(() => {
+        AsyncStorage.getItem('persist:primary')
+          .then(value => {
+            console.log('BOOM:', value)
+          })
+      })
   }
 
   render() {
