+++ App/Redux/configureStore.ts
@@ -1,15 +1,12 @@
 import { createStore, applyMiddleware } from 'redux'
-import { persistStore, persistReducer } from 'redux-persist'
+import { persistStore } from 'redux-persist'
 import createSagaMiddleware from 'redux-saga'
 import Reactotron from 'reactotron-react-native'
 
 import DebugConfig from '../Config/DebugConfig'
-import persistConfig from '../Config/ReduxPersist'
 import rootReducer from './RootReducer'
 import StartupActions from '../Redux/StartupRedux'
 import rootSaga from '../Sagas'
-
-const persistedReducer = persistReducer(persistConfig, rootReducer)
 
 export default () => {
   const sagaMonitor = DebugConfig.useReactotron
@@ -21,7 +18,7 @@ export default () => {
     ? Reactotron.createStore
     : createStore
   const store = createAppropriateStore(
-    persistedReducer,
+    rootReducer,
     applyMiddleware(sagaMiddleware)
   )
