+++ App/features/group/reducer.ts
@@ -1,5 +1,7 @@
 import { combineReducers } from 'redux'
 import { ActionType } from 'typesafe-actions'
+import { createMigrate, PersistConfig, MigrationManifest, persistReducer } from 'redux-persist'
+import { AsyncStorage } from 'react-native'
 
 import { feedReducer, FeedState } from './feed'
 import { addMessageReducer, AddMessageState } from './add-message'
@@ -17,9 +19,29 @@ export interface GroupState {
 
 export type GroupAction = ActionType<typeof actions>
 
-export default combineReducers<GroupState>({
+const migrations: MigrationManifest = {
+  0: persistedState => {
+    const state = persistedState as any
+    const { feed, addMessage, renameGroup, ...rest } = state
+    return rest
+  }
+}
+
+const persistConfig: PersistConfig = {
+  key: 'group',
+  storage: AsyncStorage,
+  version: 0,
+  whitelist: [
+    'addPhoto'
+  ],
+  migrate: createMigrate(migrations, { debug: false })
+}
+
+const reducer = combineReducers<GroupState>({
   feed: feedReducer,
   addMessage: addMessageReducer,
   addPhoto: addPhotoReducer,
   renameGroup: renameGroupReducer
 })
+
+export default persistReducer(persistConfig, reducer)
